Move IAwsS3Client to standard interface namespace LambdaS3FileZipper.Interfaces
